Categories

+++ Project ASP.Net/Repository/CategoriesRepository.cs
@@ -11,8 +11,15 @@ namespace Project_ASP.Net.Repository
         public int Delete(int id)
         {
             Category oldCategory = FindById(id);
-            db.Categories.Remove(oldCategory);
-            return db.SaveChanges();
+            if (oldCategory != null)
+            {
+                db.Categories.Remove(oldCategory);
+                return db.SaveChanges();
+            }
+            else
+            {
+                return 0;
+            }
         }
 
         public int Edit(int id, Category cate)
